fix(macos): resolve MainWindowView body evaluation hang (LUM-841)
Consolidate 15 .onReceive(NotificationCenter...) modifiers into a single
.task with async sequences, extract TopBarView as a standalone View struct,
and wrap chatContentView and overlay chain in ObservationBoundaryView to
narrow observation scope.

Part of LUM-841
